pyproject.toml
[tool.poetry]
name = "symfexit"
version = "0.1.0"
description = ""
authors = ["Jelle Besseling <jelle@pingiun.com>"]
readme = "README.md"
license = "EUPL-1.2"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
django-tailwind = "^3.4.0"
django-browser-reload = "^1.6.0"
django-constance = {extras = ["database"], version = "^3.1.0"}
fontawesomefree = "^6.2.1"
pillow = "^10.0.0"
django-tinymce = "^3.6.1"
mollie-api-python = "^3.6.0"
psycopg2 = "^2.9.9"
dj-database-url = "^2.1.0"
hashids = "^1.3.1"


[tool.poetry.group.dev.dependencies]
black = "^22.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
